requirements.txt
requests==2.28.2
PyQt6==6.5.0
PyYAML==6.0
python-dotenv==1.0.0
psutil==5.9.5
gputil==1.4.0
langchain==0.0.267
numpy==1.24.3
pillow==9.5.0 